Move to test_dataset.py
